+++ crates/aide/Cargo.toml
@@ -24,10 +24,10 @@ http = { version = "1.0.0", optional = true }
 
 axum = { version = "0.7.1", optional = true }
 axum-extra = { version = "0.9.0", optional = true }
-axum_typed_multipart = { version = "0.11.0", optional = true }
 tower-layer = { version = "0.3", optional = true }
 tower-service = { version = "0.3", optional = true }
 cfg-if = "1.0.0"
+
 
 # custom axum extractors
 serde_qs = { version = "0.12.0", optional = true }
@@ -48,7 +48,7 @@ axum-extra-cookie = ["axum", "axum-extra", "axum-extra/cookie"]
 axum-extra-cookie-private = ["axum", "axum-extra", "axum-extra/cookie-private"]
 axum-extra-form = ["axum", "axum-extra", "axum-extra/form"]
 axum-extra-query = ["axum", "axum-extra", "axum-extra/query"]
-axum-typed-multipart = ["axum_typed_multipart"]
+
 
 serde_qs = ["dep:serde_qs"]
 jwt-authorizer = ["dep:jwt-authorizer"]
